+++ src/application/Randomizer.h
@@ -3,19 +3,20 @@
 #include "../output/locationRepository/LocationRepository.h"
 #include "../output/displayManager/DisplayManager.h"
 #include "../output/itemRepository/ItemRepository.h"
-#include "../output/upgradeManager/UpgradeManager.h"
+#include "../output/characterManager/CharacterManager.h"
 #include "../output/worldStateManager/WorldStateManager.h"
 #include "../output/archipelagoMessenger/ArchipelagoMessenger.h"
 #include "structs/LocationData.h"
+#include "structs/Options.h"
 
 class Randomizer
 {
 public:
-    Randomizer(DisplayManager& displayManager, UpgradeManager& upgradeManager, WorldStateManager& menuManager,
+    Randomizer(DisplayManager& displayManager, CharacterManager& characterManager, WorldStateManager& menuManager,
                ItemRepository& itemRepository, LocationRepository& locationRepository,
-               ArchipelagoMessenger archipelagoMessenger)
+               ArchipelagoMessenger& archipelagoMessenger)
         : _displayManager(displayManager),
-          _upgradeManager(upgradeManager),
+          _characterManager(characterManager),
           _worldStateManager(menuManager),
           _itemRepository(itemRepository),
           _locationRepository(locationRepository),
@@ -34,15 +35,29 @@ public:
     void ShowStatusInformation(std::string information);
     void QueueNewMessage(std::string information);
     void OnEmblemGoalSet(int emblemGoal);
+    void SetStatingArea(StartingArea startingArea);
     void SetMissions(Characters characters, int missions);
+    void SetDeathLink(bool deathLinkActive);
+    void SetRingLink(bool ringLinkActive);
+    void SetRingLoss(RingLoss ringLoss);
+    Options GetOptions() const;
     std::vector<LifeBoxLocationData> GetLifeCapsules();
-    void OnLifeSanitySet(bool lifeSanity);
+    void ProcessDeath(const std::string& deathCause);
+    void OnPlayingFrame();
+    void OnSync();
+    void OnDeath();
+    void ProcessRings(Sint16 amount);
 
 private:
     DisplayManager& _displayManager;
-    UpgradeManager& _upgradeManager;
+    CharacterManager& _characterManager;
     WorldStateManager& _worldStateManager;
     ItemRepository& _itemRepository;
     LocationRepository& _locationRepository;
     ArchipelagoMessenger& _archipelagoMessenger;
+
+    Options _options;
+    std::string _pendingDeathCause;
+    bool _deathPending;
+    bool _ignoreNextDeath = false;
 };