+++ src/application/structs/UnlockStatus.h
@@ -3,7 +3,7 @@
 struct UnlockStatus
 {
     int currentEmblems;
-    int emblemGoal;
+    // int emblemGoal;
 
     bool sonicUnlocked;
     bool sonicLightShoesUnlocked;
@@ -29,6 +29,7 @@ struct UnlockStatus
     bool bigUnlocked;
     bool bigLifeRingUnlocked;
     bool bigPowerRodUnlocked;
+    int bigLureQuantity = 0;
     
     bool keyTrain;
     bool keyBoat;