+++ src/input/archipelago/ArchipelagoManager.h
@@ -14,7 +14,7 @@ enum ConnectionStatus
 class ArchipelagoManager
 {
 public:
-    explicit ArchipelagoManager(Randomizer& randomizer);
+    explicit ArchipelagoManager(Randomizer& randomizer, int instanceId, int64_t baseId);
 
     void Connect();
     bool IsValidSaveFile();
@@ -23,6 +23,10 @@ public:
     void OnSaveFileLoaded();
     void SetServerConfiguration(const std::string& serverIP, const std::string& playerName,
                                 const std::string& serverPassword);
+
+    std::string playerName;
+    int instanceId = 0;
+    int64_t baseId;
 
 private:
     void ManageMessages();
@@ -35,9 +39,9 @@ private:
 
     std::string _serverIP;
     std::string _serverPassword;
-    std::string _playerName;
     std::string _seedName;
 
     const float _suggestChangingConfigWaitTime = 2.5f;
     std::clock_t _connectedAt = -1;
+
 };