+++ src/input/eventDetector/EventDetector.cpp
@@ -185,7 +185,7 @@ void EventDetector::OnPlayingFrame() const
         randomizer.OnGameCompleted();
 
     //Ignore events given by the mod itself
-    if (GameMode != GameModes_Adventure_Field)
+    if (GameMode != GameModes_Adventure_Field && GameMode != GameModes_Adventure_ActionStg)
         return;
 
     bool checksFound = false;
@@ -308,4 +308,12 @@ FunctionHook<void, EntityData1*> OnExtraLife(0x4D6D40, [](EntityData1* entity)->
     }
 
     return OnExtraLife.Original(entity);
+});
+
+
+FunctionHook<void, int> onGiveLives(0x425B60, [](int lives)-> void
+{
+    if (lives == -1 && GameState == MD_GAME_FADEOUT_MISS)
+        eventDetector->randomizer.OnDeath();
+    onGiveLives.Original(lives);
 });