+++ src/output/saveFileManager/SaveFileManager.cpp
@@ -61,10 +61,11 @@ void SaveFileManager::OnSaveFileCreated()
 
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_MR_LIFEBELT));
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_MR_POWERROD));
-    // ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_BLUE));
-    // ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_GOLD));
-    // ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_RED));
-    // ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_SILVER));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_BLUE));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_GOLD));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_NORMAL));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_RED));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_RUAR_SILVER));
 
     //Train
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_TRAIN));
@@ -119,7 +120,7 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_KNUCKLES_SS_TPARK_ELEVATOR));
     ClearEventFlag(static_cast<EventFlags>(FLAG_AMY_SS_TPARK_ELEVATOR));
     ClearEventFlag(static_cast<EventFlags>(FLAG_E102_SS_TPARK_ELEVATOR));
-    // ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_SS_TPARK_ELEVATOR));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_SS_TPARK_ELEVATOR));
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_SS_ENTRANCE_CIRCUIT));
 
     //Speed Highway