+++ src/output/upgradeManager/UpgradeManager.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "../../pch.h"
-
-class UpgradeManager
-{
-public:
-    void GiveUpgrade(Upgrades upgrade);
-    void RemoveUpgrade(Upgrades upgrade);
-};